Factor out form-submitter.tsx.
